Locking Down DeleteItem Permissions

+++ backend/src/resolvers/Mutation.js
@@ -48,9 +48,15 @@ const Mutations = {
     const where = { id: args.id }
 
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`)
 
     // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId 
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You not allowed')
+    }
     // 3. Delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
